Move metric trigger tests to existing test file
- Remove test_metric_trigger_deserialization.py
- Add tests to test_automations_server_compatibility.py where they belong
